chapter5/lesson1/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
Pillow>=10.0.0 